Convert project to Maven structure
