+++ README.md
@@ -1,2 +1,9 @@
 # EW2_LIBRARY
 A Java-based project developed during the Ervaringsweek 2 module. The repository contains exercises, examples, and implementations covering core object-oriented programming concepts, UML design, and integration with libraries.
+
+## Build & Run
+
+```bash
+mvn -q clean test
+mvn -q exec:java -Dexec.mainClass="be.intecbrussel.library.app.Main"
+```
