--- be.intecbrussel/src/model/BookType.java
@@ -1,4 +1,0 @@
-package model;
-
-public enum BookType {
-}
