--- be.intecbrussel/src/model/Library.java
@@ -1,4 +1,0 @@
-package model;
-
-public class Library {
-}
