--- be.intecbrussel/src/model/LoanStatus.java
@@ -1,4 +1,0 @@
-package model;
-
-public enum LoanStatus {
-}
