--- be.intecbrussel/src/repository/BookRepositoryImpl.java
@@ -1,4 +1,0 @@
-package repository;
-
-public class BookRepositoryImpl {
-}
