--- be.intecbrussel/src/repository/IBookRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public interface IBookRepository {
-}
