--- be.intecbrussel/src/repository/ILoanRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public interface ILoanRepository {
-}
