--- be.intecbrussel/src/repository/IMemberRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public interface IMemberRepository {
-}
