--- be.intecbrussel/src/repository/LoanRepositoryImpl.java
@@ -1,4 +1,0 @@
-package repository;
-
-public class LoanRepositoryImpl {
-}
