--- be.intecbrussel/src/repository/MemberRepositoryImpl.java
@@ -1,4 +1,0 @@
-package repository;
-
-public class MemberRepositoryImpl {
-}
