--- be.intecbrussel/src/service/LibraryService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class LibraryService {
-}
