+++ src/main/java/be/intecbrussel/library/model/BookType.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.model;
+
+public enum BookType {
+}
