+++ src/main/java/be/intecbrussel/library/model/Library.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.model;
+
+public class Library {
+}
