+++ src/main/java/be/intecbrussel/library/model/Loan.java
@@ -1,4 +1,4 @@
-package model;
+package be.intecbrussel.library.model;
 
 import java.time.LocalDate;
