+++ src/main/java/be/intecbrussel/library/model/LoanStatus.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.model;
+
+public enum LoanStatus {
+}
