+++ src/main/java/be/intecbrussel/library/model/Member.java
@@ -1,4 +1,4 @@
-package model;
+package be.intecbrussel.library.model;
 
 
 public class Member {
