+++ src/main/java/be/intecbrussel/library/repository/BookRepositoryImpl.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.repository;
+
+public class BookRepositoryImpl {
+}
