+++ src/main/java/be/intecbrussel/library/repository/IBookRepository.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.repository;
+
+public interface IBookRepository {
+}
