+++ src/main/java/be/intecbrussel/library/repository/ILoanRepository.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.repository;
+
+public interface ILoanRepository {
+}
