+++ src/main/java/be/intecbrussel/library/service/LibraryService.java
@@ -1,0 +1,4 @@
+package be.intecbrussel.library.service;
+
+public class LibraryService {
+}
